1、完成了充值接口

+++ src/main/java/com/example/rv/pojo/Users.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Pattern;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.time.LocalDateTime;
 
@@ -20,8 +21,8 @@ public class Users {
     @NotNull
     @Pattern(regexp = "^$|^1[3-9]\\d{9}$", message = "手机号码格式错误")
     private String userPhoneNumber;
-    private float userBalance;
-    private float userPoints;
+    private BigDecimal userBalance;
+    private BigDecimal userPoints;
     private Timestamp userCreatedAt;
     private Timestamp userUpdatedAt;
 }
